+++ src/serial.f90
@@ -156,7 +156,7 @@ character(1), intent(in) :: mode
 character(*), intent(in) :: filename
 integer, intent(in) :: m, o, mpio
 logical, intent(in) :: verb
-real :: f2(1,size(f1))
+integer :: f2(1,size(f1))
 integer :: i
 if ( mode == 'w' ) f2(1,:) = f1
 call fiio2( fh, f2, mode, filename, m, o, verb )
